refactor: extract navbar variants to sub-folder
