+++ frontend/src/components/layout/Navbar.tsx
@@ -1,8 +1,8 @@
 import { NavLink } from 'react-router-dom';
 import TournamentSelector from './TournamentSelector';
-import DesktopNav from './DesktopNav';
-import CompactNav from './CompactNav';
-import QuickNav from './QuickNav';
+import DesktopNav from './navbar/DesktopNav';
+import CompactNav from './navbar/CompactNav';
+import QuickNav from './navbar/QuickNav';
 import ThemeToggle from './ThemeToggle';
 
 import { PRIMARY_NAV_ITEMS, NAV_ITEMS } from '@/constants/navigation';
