+++ frontend/src/components/layout/navbar/CompactNav.tsx
