+++ frontend/src/components/layout/navbar/DesktopNav.tsx
