+++ frontend/src/components/layout/navbar/QuickNav.tsx
